Begin putting student home in grid form

+++ public/javascripts/Controllers/studentHomeCtrl.js
@@ -1,20 +1,26 @@
 angular.module('biologyGraphingApp').controller('StudentHomeCtrl', ['$scope', 'documents', 'sections', 'assignments', 'auth',
 function($scope, documents, sections, assignments, auth) {
 	$scope.documents = documents.documents;
-	$scope.newDocTitle = '';
 	$scope.sections = sections.sections;
 	$scope.selections = {assignment: null, document: null};
 	$scope.isLoggedIn = auth.isLoggedIn;
 
   $scope.addDocument = function() {
-    if ($scope.newDocTitle === '') { return; }
+		var newDocTitle = prompt("Enter a Title for your new Document.", "");
+    if (newDocTitle === '' || newDocTitle == null) { return; }
     var graph = { nodes: [], edges: [], undoStack: [] };
-    documents.addDocument($scope.newDocTitle, graph);
-    $scope.newDocTitle = '';
+    documents.addDocument(newDocTitle, graph);
   };
 
   $scope.deleteDocument = function(document) {
-    documents.deleteDocument(document);
+		if (document.status == 'unsubmitted') {
+			if (confirm("Are you sure you want to delete " + document.title + "?")) {
+	    	documents.deleteDocument(document);
+	    }
+		}
+		else {
+			alert("Cannot delete " + document.title + " because it has been submitted to " + document.submittedTo.title + ".");
+		}
   };
 
 	$scope.addSubmission = function() {
